fer2013/requirements.txt
torch
torchvision
numpy
tqdm
scikit-learn
matplotlib
seaborn
torchinfo
PyYAML
